src/app/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
pydantic_settings==2.3.4
sqlmodel==0.0.19
psycopg==3.2.1
psycopg-binary==3.2.1
pandas==2.2.2
scikit-learn==1.5.1
pyarrow==16.1.0
uuid==1.30
pika==1.3.2
passlib==1.7.4
python-jose==3.3.0
pytest==8.2.2
requests==2.32.3